==> notebooks/3_A_understand_data.ipynb <==
# Understand data
# Objective: Conduct a general inspection on the released datasets

# importing the require libraries to run the code in the this notebook
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql.types import StringType
import re
import pandas as pd

# running spark
spark = (
    SparkSession.builder.appName("preprocessing of taxi data")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)


# reading in the require dataset
merchant_sdf = spark.read.parquet("../data/tables/tbl_merchants.parquet")
consumer_detail_sdf = spark.read.parquet("../data/tables/consumer_user_details.parquet")
consumer_sdf = spark.read.option("header",True) \
                         .option("inferSchema",True) \
                         .options(delimiter='|') \
                         .csv("../data/tables/tbl_consumer.csv")
transaction_28_sdf = spark.read.parquet("../data/tables/transactions_20210228_20210827_snapshot")
transaction_82_sdf = spark.read.parquet("../data/tables/transactions_20210828_20220227_snapshot/")

# Transaction dataset
- should be our main focus in terms on analysing the number of transactions related to each customer and each merchant
- could be joined to consumer_detail by using user_id and consumer_id with the help of tbl_consumer
- join merchant on merchant abn (should be unique by personal knowledge)

transaction_28_sdf.orderBy(F.column("order_datetime"))

# Merchant Dataset

# quick look ar a random row
merchant_sdf.where(F.col("merchant_abn") == "20692490685")

merchant_sdf.select(F.col("tags")).take(1)

# Consumer Key (conversion table)

consumer_detail_sdf

# Consumer Dataset

consumer_sdf

consumer_detail_sdf.printSchema()

# Check for null values in datasets (except merchant)

transaction_82_sdf.select(F.col("user_id")).count()

consumer_missing = consumer_sdf.select([F.count(F.when(F.col(c).contains('None') | \
                                                       F.col(c).contains('NULL') | \
                                                      (F.col(c) == '' ) | \
                                                       F.col(c).isNull() | \
                                                       F.isnan(c), c 
                                                      )).alias(c)
                                                       for c in consumer_sdf.columns])
consumer_missing.show()


consumer_detail_missing = consumer_detail_sdf.select([F.count(F.when(F.col(c).contains('None') | \
                                                       F.col(c).contains('NULL') | \
                                                      (F.col(c) == '' ) | \
                                                       F.col(c).isNull() | \
                                                       F.isnan(c), c 
                                                      )).alias(c)
                                                       for c in consumer_detail_sdf.columns])
consumer_detail_missing.show()

transact_col = ['user_id', 'merchant_abn', 'dollar_value', 'order_id']
transaction_missing = transaction_28_sdf.select([F.count(F.when(F.col(c).contains('None') | \
                                                       F.col(c).contains('NULL') | \
                                                      (F.col(c) == '' ) | \
                                                       F.col(c).isNull() | \
                                                       F.isnan(c), c 
                                                      )).alias(c)
                                                       for c in transact_col])
transaction_missing.show()

transact_col = ['user_id', 'merchant_abn', 'dollar_value', 'order_id']
transaction_missing = transaction_82_sdf.select([F.count(F.when(F.col(c).contains('None') | \
                                                       F.col(c).contains('NULL') | \
                                                      (F.col(c) == '' ) | \
                                                       F.col(c).isNull() | \
                                                       F.isnan(c), c 
                                                      )).alias(c)
                                                       for c in transact_col])
transaction_missing.show()

# Fixed tags in Merchant
The tags in merchant are all grouped in the same column lets, split it to each respective columns

# regex to split the tags
split_col = F.split(merchant_sdf['tags'], '\]|\)', 3)
merchant_sdf = merchant_sdf.withColumn('prod_desc', split_col.getItem(0))
merchant_sdf = merchant_sdf.withColumn('revenue_level', split_col.getItem(1))
merchant_sdf = merchant_sdf.withColumn('take_rate', split_col.getItem(2))

merchant_sdf

merchant_sdf.select(F.col("take_rate")).take(5)

@F.udf(returnType=StringType())
def clean_str(str):
    """ cleans string after extractions from TAGS columns, 
        remove brakcets.

    Args:
        str (str): 

    Returns:
        str: lowercase, with no leading or trailing parenthesis 
    """
    str = str.lstrip(' ,([')
    str = str.rstrip(')]')
    str = str.lower()
    return str

# cleaning the tags
merchant_sdf = merchant_sdf.withColumn('prod_desc', clean_str(F.col('prod_desc')))
merchant_sdf = merchant_sdf.withColumn('take_rate', clean_str(F.col('take_rate')))
merchant_sdf = merchant_sdf.withColumn('revenue_level', clean_str(F.col('revenue_level')))
merchant_sdf = merchant_sdf.withColumn('take_rate', F.regexp_extract('take_rate', '\\d*\\.\\d', 0))
merchant_sdf = merchant_sdf.withColumn('take_rate', F.col('take_rate').cast('double'))
merchant_sdf = merchant_sdf.drop("tags")

merchant_sdf

## Check for missing

merchant_missing = merchant_sdf.select([F.count(F.when(F.col(c).contains('None') | \
                                                       F.col(c).contains('NULL') | \
                                                      (F.col(c) == '' ) | \
                                                       F.col(c).isNull() | \
                                                       F.isnan(c), c 
                                                      )).alias(c)
                                                       for c in merchant_sdf.columns])
merchant_missing.show()
