==> notebooks/6_R_C_fraud_inspect.ipynb <==
# Inspect Fraud Data

Purpose
- Have a look at what the teaching staff's has given us and come up with a plan to deal with this fraud problem

# import libraries
import pandas as pd
from pyspark.sql import SparkSession
from pyspark.sql import functions as F

# setup spark
spark = (
    SparkSession.builder.appName("aggregate data for first 3 final model variables")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)

merchant_fraud = pd.read_csv('../data/tables/merchant_fraud_probability.csv')

data1 = spark.read.parquet("../data/tables/transactions_20210228_20210827_snapshot/")
data2 = spark.read.parquet("../data/tables/transactions_20210828_20220227_snapshot/")
data3 = spark.read.parquet("../data/tables/transactions_20220228_20220828_snapshot/")

merchant_info = spark.read.parquet("../data/tables/tbl_merchants.parquet")
merchant_info_df = merchant_info.toPandas()

merchant_info_df[merchant_info_df['merchant_abn'] == 19492220327]

data = data1.union(data2)
data = data.union(data3)

consumer_fraud = pd.read_csv('../data/tables/consumer_fraud_probability.csv', )

data[(data['user_id'] == 6228) & (data['order_datetime']=='2021-12-19')]

data[(data['merchant_abn'] == 19492220327) & (data['order_datetime']=='2021-11-28')]

consumer_fraud

consumer_fraud.describe()

consumer_fraud.nunique()

merchant_fraud

merchant_fraud.nunique()

merchant_fraud.describe()
