==> notebooks/7_R_C_fraud_data_clean.ipynb <==
# Create Training Data for Fraud Predictor Models

Purpose:
- Clean data from raw transactions into aggregated daily behaviour (i.e. daily transaction average amount, daily number of transaction, daily number of distinct customers etc)

# import libraries
import pandas as pd
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from statistics import mean, stdev, pstdev
import os

# setup spark
spark = (
    SparkSession.builder.appName("aggregate data for first 3 final model variables")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)

# create directory persona if does not exist
dirs_to_create = ['../data/curated/fraud', '../data/curated/fraud/input', '../data/curated/fraud/output']

def create_dirs(dirs_to_create):
    # check if it exists as it makedir will raise an error if it does exist
    for dir_to_create in dirs_to_create:
        if not os.path.exists(dir_to_create):
            os.makedirs(dir_to_create)

create_dirs(dirs_to_create)

## Read in files

# import transactions data
data1 = spark.read.parquet("../data/tables/transactions_20210228_20210827_snapshot/")
data2 = spark.read.parquet("../data/tables/transactions_20210828_20220227_snapshot/")
data3 = spark.read.parquet("../data/tables/transactions_20220228_20220828_snapshot/")

data = data1.union(data2)
data = data.union(data3)

# read both fraud data
merchant_fraud = spark.read.option("header",True).csv('../data/tables/merchant_fraud_probability.csv', )
consumer_fraud = spark.read.option("header",True).csv('../data/tables/consumer_fraud_probability.csv', )

# join up the fraud data with merchant data so every transaction made on the day with fraud label is linked
merchant_fraud_join_data = merchant_fraud.join(data, [merchant_fraud.merchant_abn == data.merchant_abn, \
    merchant_fraud.order_datetime == data.order_datetime], 'left')

consumer_fraud_join_data = consumer_fraud.join(data, [consumer_fraud.user_id == data.user_id, \
    consumer_fraud.order_datetime == data.order_datetime], 'left')

# create pandas dataframe
merchant_fraud_join_data_df = merchant_fraud_join_data.toPandas()
consumer_fraud_join_data_df = consumer_fraud_join_data.toPandas()

# drop duplicate columns from previous join
merchant_fraud_join_data_df = merchant_fraud_join_data_df.iloc[:, 2:]
consumer_fraud_join_data_df = consumer_fraud_join_data_df.iloc[:, 2:]

# create exxtra column for taking aggregation
merchant_fraud_join_data_df['dollar_value2'] = merchant_fraud_join_data_df['dollar_value']
consumer_fraud_join_data_df['dollar_value2'] = consumer_fraud_join_data_df['dollar_value']

# take aggregation
merchant_fraud_instance_agg = merchant_fraud_join_data_df.groupby(['merchant_abn', 'order_datetime']).agg({'dollar_value': mean, \
    'dollar_value2': pstdev, 'order_id':'count', 'user_id':'nunique', 'fraud_probability': 'first'})

consumer_fraud_instance_agg = consumer_fraud_join_data_df.groupby(['user_id', 'order_datetime']).agg({'dollar_value': mean, \
    'dollar_value2': pstdev, 'order_id':'count', 'merchant_abn':'nunique', 'fraud_probability': 'first'})

# previously used population standard deviation to avoid dividng by 0 error. 
# Now use liottle trick to get sample sd and setting samples with only one obs to 0
tmp = merchant_fraud_instance_agg['order_id']
tmp2 = [x-1 if x != 1 else 1 for x in tmp ]
merchant_fraud_instance_agg['count**'] = tmp2
merchant_fraud_instance_agg['sd_transact'] = merchant_fraud_instance_agg['dollar_value2']* \
    merchant_fraud_instance_agg['order_id']/merchant_fraud_instance_agg['count**']
merchant_fraud_instance_agg = merchant_fraud_instance_agg.drop(['dollar_value2', 'count**'], axis=1)

tmp = consumer_fraud_instance_agg['order_id']
tmp2 = [x-1 if x != 1 else 1 for x in tmp ]
consumer_fraud_instance_agg['count**'] = tmp2
consumer_fraud_instance_agg['sd_transact'] = consumer_fraud_instance_agg['dollar_value2']* \
    consumer_fraud_instance_agg['order_id']/consumer_fraud_instance_agg['count**']
consumer_fraud_instance_agg = consumer_fraud_instance_agg.drop(['dollar_value2', 'count**'], axis=1)

# rename for more accurate feature descriptions
merchant_fraud_instance_agg = merchant_fraud_instance_agg.rename(columns = {'order_id': '#daily_orders', 'dollar_value': 'transact_amount_perOrder', 'user_id': '#distinct_customers', 'sd_transact': 'transact_amount_perOrder_sd'})
consumer_fraud_instance_agg = consumer_fraud_instance_agg.rename(columns = {'order_id': '#daily_orders', 'dollar_value': 'transact_amount_perOrder', 'merchant_abn': '#distinct_merchants', 'sd_transact': 'transact_amount_perOrder_sd'})

# change it into correct data type of float
merchant_fraud_instance_agg['fraud_probability'] = merchant_fraud_instance_agg['fraud_probability'].astype(float)
consumer_fraud_instance_agg['fraud_probability'] = consumer_fraud_instance_agg['fraud_probability'].astype(float)

## Analysis

merchant_fraud_instance_agg.describe()

consumer_fraud_instance_agg.describe()

consumer_fraud_instance_agg

# Get all merchant/consumer's daily activity - to take ratio (more feature engineering)

# First aggregate by day
full_merchant_daily = data.groupBy("merchant_abn", "order_datetime").agg(F.countDistinct("user_id").alias('distinct_users'),
                                                     F.count("order_id").alias('count'),
                                                     F.mean("dollar_value").alias('mean transact'),
                                                     F.stddev("dollar_value").alias('stdev transact'))

# Then aggregate by unit
full_merchant_daily_agg = full_merchant_daily.groupBy("merchant_abn")\
    .agg(F.mean('distinct_users').alias('mean_#distinct_customers'), \
        F.mean('count').alias('mean_#daily_orders'),\
        F.mean("mean transact").alias('mean_transact_amount_perOrder'),\
        F.mean('stdev transact').alias('mean_transact_amount_perOrder_sd'))

# turn into Pandas dataframe
full_merchant_daily_agg_df = full_merchant_daily_agg.toPandas()

# First aggregate by day
full_consumer_daily = data.groupBy("user_id", "order_datetime").agg(F.countDistinct("merchant_abn").alias('distinct_merchants'),
                                                     F.count("order_id").alias('count'),
                                                     F.mean("dollar_value").alias('mean transact'),
                                                     F.stddev("dollar_value").alias('stdev transact'))

# Then aggregate by unit
full_consumer_daily_agg = full_consumer_daily.groupBy("user_id")\
    .agg(F.mean('distinct_merchants').alias('mean_#distinct_merchants'), \
        F.mean('count').alias('mean_#daily_orders'),\
        F.mean("mean transact").alias('mean_transact_amount_perOrder'),\
        F.mean('stdev transact').alias('mean_transact_amount_perOrder_sd'))

# turn into Pandas dataframe
full_consumer_daily_agg_df = full_consumer_daily_agg.toPandas()                                      

# create ratio variables after merging fraud set with all merchant's individual averagehistorical activity
merchant_fraud_df = merchant_fraud_instance_agg.merge(full_merchant_daily_agg_df, on = 'merchant_abn', how = 'left')

merchant_fraud_df['#daily_orders_ratio'] = merchant_fraud_df['#daily_orders']/merchant_fraud_df['mean_#daily_orders']
merchant_fraud_df['#distinct_customers_ratio'] = merchant_fraud_df['#distinct_customers']/merchant_fraud_df['mean_#distinct_customers']
merchant_fraud_df['transact_amount_perOrder_sd_ratio'] = merchant_fraud_df['transact_amount_perOrder_sd']/merchant_fraud_df['mean_transact_amount_perOrder_sd']
merchant_fraud_df['transact_amount_perOrder_ratio'] = merchant_fraud_df['transact_amount_perOrder']/merchant_fraud_df['mean_transact_amount_perOrder']

merchant_fraud_df = merchant_fraud_df.drop(['mean_#daily_orders', 'mean_#distinct_customers', 'mean_transact_amount_perOrder_sd', \
    'mean_transact_amount_perOrder'], axis=1)

merchant_fraud_df = merchant_fraud_df.fillna(0)

# create ratio variables after merging fraud set with all user's average historical activity 
consumer_fraud_df = consumer_fraud_instance_agg.merge(full_consumer_daily_agg_df, on = 'user_id', how = 'left')

consumer_fraud_df['#daily_orders_ratio'] = consumer_fraud_df['#daily_orders']/consumer_fraud_df['mean_#daily_orders']
consumer_fraud_df['#distinct_merchants_ratio'] = consumer_fraud_df['#distinct_merchants']/consumer_fraud_df['mean_#distinct_merchants']
consumer_fraud_df['transact_amount_perOrder_sd_ratio'] = consumer_fraud_df['transact_amount_perOrder_sd']/consumer_fraud_df['mean_transact_amount_perOrder_sd']
consumer_fraud_df['transact_amount_perOrder_ratio'] = consumer_fraud_df['transact_amount_perOrder']/consumer_fraud_df['mean_transact_amount_perOrder']

consumer_fraud_df = consumer_fraud_df.drop(['mean_#daily_orders', 'mean_#distinct_merchants', 'mean_transact_amount_perOrder_sd', 'mean_transact_amount_perOrder'], axis=1)

consumer_fraud_df = consumer_fraud_df.fillna(0)

consumer_fraud_df.describe()

merchant_fraud_df.describe()

# Export
merchant_fraud_df.to_csv('../data/curated/fraud/input/merchant_fraud_model_building_data.csv', index = False)

consumer_fraud_df.to_csv('../data/curated/fraud/input/consumer_fraud_model_building_data.csv', index = False)

# Aggregate data of all consumers for running fraud detection on (only take consumer data because failed to build model for merchants)

# get all consumer data which will be used in 
full_consumer_daily_df = full_consumer_daily.toPandas()

# create ratio variables after merging fraud set with all user's average historical activity 
full_consumer_fraud_df = full_consumer_daily_df.merge(full_consumer_daily_agg_df, on = 'user_id', how = 'left')

full_consumer_fraud_df['#daily_orders_ratio'] = full_consumer_fraud_df['count']/full_consumer_fraud_df['mean_#daily_orders']
full_consumer_fraud_df['#distinct_merchants_ratio'] = full_consumer_fraud_df['distinct_merchants']/full_consumer_fraud_df['mean_#distinct_merchants']
full_consumer_fraud_df['transact_amount_perOrder_sd_ratio'] = full_consumer_fraud_df['stdev transact']/full_consumer_fraud_df['mean_transact_amount_perOrder_sd']
full_consumer_fraud_df['transact_amount_perOrder_ratio'] = full_consumer_fraud_df['mean transact']/full_consumer_fraud_df['mean_transact_amount_perOrder']

full_consumer_fraud_df = full_consumer_fraud_df.drop(['mean_#daily_orders', 'mean_#distinct_merchants', 'mean_transact_amount_perOrder_sd', 'mean_transact_amount_perOrder'], axis=1)

full_consumer_fraud_df = full_consumer_fraud_df.fillna(0)

full_consumer_fraud_df = full_consumer_fraud_df.rename(columns={'distinct_merchants': 'mean_#distinct_merchants', 'count': 'mean_#daily_orders',\
    'mean transact': 'mean_transact_amount_perOrder', 'stdev transact': 'mean_transact_amount_perOrder_sd'})

full_consumer_fraud_df.to_csv('../data/curated/fraud/input/full_consumer_behaviour_data_for_fraud_prediction.csv')
