==> notebooks/11_R_C_consumer runthrough.ipynb <==
# Predict Fraud and delete fraud transactions

Purpose:
- run predictions over all dataset and drop transactions that surpass threshold

- also here we determine what threshold to mark consumer-day as fraud

import pandas as pd
import pickle
from sklearn.ensemble import RandomForestRegressor

# Get the predicted Fraud labelled (for dropping once joined)

# load the rf
with open('../models/consumer_fraud_rfr.pickle', 'rb') as g:
    rfr = pickle.load(g)

# load the full data, and prepare it to undertake prediction
data = pd.read_csv('../data/curated/fraud/input/full_consumer_behaviour_data_for_fraud_prediction.csv')

data = data.rename(columns = {'mean_#distinct_merchants': '#distinct_merchants', \
        'mean_#daily_orders': '#daily_orders', 'mean_transact_amount_perOrder': 'transact_amount_perOrder',\
        'mean_transact_amount_perOrder_sd': 'transact_amount_perOrder_sd'})

X = data[['transact_amount_perOrder_ratio', 'transact_amount_perOrder',
    'transact_amount_perOrder_sd', 'transact_amount_perOrder_sd_ratio']]

# make prediction
prediction = rfr.predict(X)

df = pd.DataFrame({'prediction': prediction})

Analysis

df.describe()[1:]

# inspect how many % of transactions get dropped off for each cutoff threshold rate
# it was here where it was decided to use 20 as our threshold
for i in range(10, 85, 5):
    print(i, ':', len(df[df['prediction']>i])/len(df))

# create fraud label
data['fraud rate'] = prediction
final_fraud_prediction = data[['user_id', 'order_datetime', 'fraud rate']]
final_fraud_prediction['fraud'] = final_fraud_prediction['fraud rate'].apply(lambda x:1 if x >= 20 else 0)

final_fraud_prediction.to_csv('../data/curated/fraud/output/final_fraud_prediction.csv', index=False)

# Get the actual fraud labelled (for dropping once joined)

consumer_fraud_df = pd.read_csv('../data/tables/consumer_fraud_probability.csv')

# find threshold that drops 0.001 of labelled transactions
consumer_fraud_df.quantile(0.999)

# create labels
consumer_fraud_df['fraud'] = consumer_fraud_df['fraud_probability'].apply(lambda x:1 if x >= 83 else 0)
consumer_fraud_df

consumer_fraud_df.to_csv('../data/curated/fraud/output/final_fraud_real.csv', index=False)

# Join together and drop fraud rows

# import libraries
import pandas as pd
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from statistics import mean, stdev, pstdev
import os

# setup spark
spark = (
    SparkSession.builder.appName("aggregate data for first 3 final model variables")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)

# import transactions data
data1 = spark.read.parquet("../data/tables/transactions_20210228_20210827_snapshot/")
data2 = spark.read.parquet("../data/tables/transactions_20210828_20220227_snapshot/")
data3 = spark.read.parquet("../data/tables/transactions_20220228_20220828_snapshot/")

data = data1.union(data2)
data = data.union(data3)

# read both fraud data
final_fraud_prediction = spark.read.option("header",True).csv('../data/curated/fraud/output/final_fraud_prediction.csv', )
final_fraud_real = spark.read.option("header",True).csv('../data/curated/fraud/output/final_fraud_real.csv', )

# rename columns before join
final_fraud_real = final_fraud_real.withColumnRenamed("user_id", "user_id3")\
    .withColumnRenamed("order_datetime", "order_datetime3").withColumnRenamed("fraud", "fraud3")

final_fraud_prediction = final_fraud_prediction.withColumnRenamed("user_id", "user_id2")\
    .withColumnRenamed("order_datetime", "order_datetime2").withColumnRenamed("fraud", "fraud2")

# join up
data_withFraud = final_fraud_prediction.join(data, [final_fraud_prediction.user_id2 == data.user_id, \
    final_fraud_prediction.order_datetime2 == data.order_datetime], 'outer')

data_withFraud = data_withFraud.join(final_fraud_real, [data_withFraud.user_id == final_fraud_real.user_id3, \
    data_withFraud.order_datetime == final_fraud_real.order_datetime3], 'left').fillna(0)

data_withFraud.head()

# drop fraud (picked up by prediction)
data_withoutFraud = data_withFraud.filter(F.col('fraud2') == 0)


data_withoutFraud.count()

# drop fraud (picked up by observation)
data_withoutFraud = data_withoutFraud.filter((F.col('fraud3') == 0) | (F.col('fraud3').isNull()))

# realise that predicted fraud includes all the observed fraud, so this is good emperical evidence that our work is good
data_withoutFraud.count()

data_withoutFraud_correctDates = data_withoutFraud.filter(data_withoutFraud["order_datetime"] >= F.lit('2021-03-01')) \
       .filter(data_withoutFraud["order_datetime"] <= F.lit('2022-08-14'))

data_withoutFraud_correctDates.count()

data_withoutFraud_correctDates = data_withoutFraud_correctDates.select(['user_id', 'merchant_abn', 'dollar_value', 'order_id', 'order_datetime'])

data_withoutFraud_correctDates.write.mode('overwrite').parquet('../data/curated/fraud/output/transactions_withoutfraud')
