==> notebooks/13_H_C_A_cluster_data_prep.ipynb <==
# Data preparation for clustering method
# Objective: Generate features to a dataset for take rate imputation through clustering

# importing the require libraries to run the code in the this notebook
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql import DataFrameStatFunctions as statFunc

# create directory persona if does not exist
import os

dirs_to_create = ['../data/curated/clusters', '../data/curated/clusters/input', '../data/curated/clusters/output']

def create_dirs(dirs_to_create):
    # check if it exists as it makedir will raise an error if it does exist
    for dir_to_create in dirs_to_create:
        if not os.path.exists(dir_to_create):
            os.makedirs(dir_to_create)

create_dirs(dirs_to_create)

# running spark
spark = (
    SparkSession.builder.appName("preprocessing of taxi data")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)

# read in the require dataset
transaction_sdf = spark.read.parquet("../data/curated/fraud/output/transactions_withoutfraud")
transaction_sdf

## Perform monthly aggregation to obtain monthly number of orders, mean revenue, and standard deviation of revenue

monthly_agg = transaction_sdf.groupBy(
                                      "merchant_abn",
                                      F.year("order_datetime"),
                                      F.month("order_datetime")
                                      ) \
                                      .agg(
                                        F.count("order_id").alias("number_of_order"),
                                        F.sum("dollar_value").alias("total_dollar_value")
                                      ) \
                                      .groupBy("merchant_abn") \
                                      .agg(
                                        F.mean("number_of_order").alias("monthly_average_number_of_orders"),
                                        F.mean("total_dollar_value").alias("monthly_average_bnpl_revenue"),
                                        F.stddev("total_dollar_value").alias("stddev_of_monthly_average_bnpl_revenue")
                                      )

Get aggregated transaction dataset with engineered features, this dataset are the missing merchants that requires imputation

# find all unknown merchants by looking for merchants with only ABN but not name, revenue level, and take rate.
# And conduct feature engineering for total number of customers, standard deviation of total dollar value and mean of dollar value.
# and join with the monthly features to transaction sdf.


agg_transaction_pred_sdf = transaction_sdf.filter(F.col("merchant_name").isNull()) \
                                          .groupBy("merchant_abn") \
                                          .agg(
                                               F.countDistinct("consumer_id").alias("total_number_of_distinct_customers"),
                                               F.stddev("dollar_value").alias("stddev_of_dorllar_value"),
                                               F.mean("dollar_value").alias("average_dollar_value"),
                                               F.percentile_approx("dollar_value", 0.5).alias("median_dollar_value")
                                               ) \
                                          .join(
                                            monthly_agg,
                                            ['merchant_abn', ],
                                            "left"
                                          ).toPandas().to_csv("../data/curated/clusters/input/agg_transaction_pred.csv", index=False)                 

Get aggregated transaction dataset with engineered features, this dataset are the known merchants with information present and is use to create clusters.

# find all known merchants by looking for merchants with known merchant information
# And conduct feature engineering for total number of customers, standard deviation of total dollar value and mean of dollar value.
# and join with the monthly features to transaction sdf.
agg_transaction_train_sdf = transaction_sdf.filter(F.col("merchant_name").isNotNull()) \
                                          .groupBy("merchant_abn") \
                                          .agg(
                                               F.countDistinct("consumer_id").alias("total_number_of_distinct_customers"),
                                               F.stddev("dollar_value").alias("stddev_of_dorllar_value"),
                                               F.mean("dollar_value").alias("average_dollar_value"),
                                               F.percentile_approx("dollar_value", 0.5).alias("median_dollar_value"),
                                               F.mean("take_rate").alias("take_rate")
                                               ) \
                                          .join(
                                            monthly_agg,
                                            ['merchant_abn', ],
                                            "left"
                                          ).toPandas().to_csv("../data/curated/clusters/input/agg_transaction_train.csv", index=False)
