==> notebooks/15_C_tally_merchant_cluster.ipynb <==
# Tally of Merchant Clusters

import pandas as pd
pd.options.display.max_colwidth = 200

merchant_details = pd.read_parquet("../data/tables/tbl_merchants.parquet").reset_index('merchant_abn')
merchant_clusters = pd.read_csv("../data/curated/clusters/output/merchant_clusters.csv")

display(merchant_details.head(3))
display(merchant_clusters.head(3))

print('# distinct abn in merchant_clusters:', len(merchant_clusters.merchant_abn.unique()))
print('# distinct abn in merchant_details:', len(merchant_details.merchant_abn.unique()))

new_merchant = sum([1 for abn in set(merchant_clusters.merchant_abn.unique()) if abn not in set(merchant_details.merchant_abn.unique())])
print('# new merchant appears in merchant_clusters:', new_merchant)

## Extract the store categories

**Clean the tags**

# split the product description, revenue level and take rate
merchant_details[['prod_desc','revenue_level','take_rate']] = merchant_details.tags.str.split("\],|\), ",expand=True)
# only keep the product description
merchant_details = merchant_details.drop(columns=['tags','revenue_level','take_rate', 'name'])
# clean the strings
merchant_details.prod_desc = merchant_details.prod_desc.map(lambda x: x.lstrip(' ,([').rstrip(' )]').lower().replace('  ', ' '))
display(merchant_details.head())

**We only have 25 distinct product descriptions**

display(merchant_details.describe())
display(set(merchant_details.prod_desc.values))

## Build the Tally!

**Inner join the product description and clustered labels**

prod_label_merged = merchant_clusters.merge(merchant_details, on='merchant_abn', how='inner')\
    .drop(columns=['merchant_abn'])
display(prod_label_merged.head(3))

**Create an empty table**

my_tally = pd.DataFrame(index=set(merchant_details.prod_desc.unique()), columns=['cluster_' + str(cluster) for cluster in set(merchant_clusters.label.unique())]).fillna(0)
my_tally.index.name = 'product_description'
#display(my_tally.head(3))

**Fill the frequency in!**

for cluster in set(merchant_clusters.label.unique()):
    for prod_desc in prod_label_merged[prod_label_merged.label == cluster].prod_desc.values:
        my_tally.loc[prod_desc, 'cluster_'+str(cluster)] += 1

my_tally['sum'] = my_tally.sum(axis=1)    
display(my_tally)

my_tally.sum()

for i in range(3):
    display(my_tally.sort_values([f'cluster_{i}'], ascending=False)[f'cluster_{i}'].head(5))

## Take a look at the percentages

**By product descriptions**

byProd_tally = my_tally.iloc[:,0:3].copy()
for col in byProd_tally.columns:
    byProd_tally[col] = byProd_tally[col]/my_tally['sum']


byProd_tally

for i in range(3):
    display(byProd_tally.sort_values([f'cluster_{i}'], ascending=False)[f'cluster_{i}'].head(5))

**By cluster labels**

byCluster_tally = my_tally.iloc[:,0:3].copy()
for col in byCluster_tally.columns:
    byCluster_tally[col] = byCluster_tally[col]/my_tally[col].sum()

byCluster_tally

for i in range(3):
    display(byCluster_tally.sort_values([f'cluster_{i}'], ascending=False)[f'cluster_{i}'].head(5))

## Save to CSV

my_tally.to_csv('../data/curated/clusters/output/tally_merchant_cluster_prod_desc.csv')
