==> notebooks/17_R_combine_take_rate_and_imputed_take_rate.ipynb <==
# Combine take rate and imputed take rate into one file

Purpose:
- take cluster output as input and return full take rate for all merchants

import pandas as pd

# read in data
clusters = pd.read_csv('../data/curated/clusters/output/merchant_clusters.csv')
cluster_take_rate = pd.read_csv('../data/curated/clusters/output/take_rate.csv')
merchant_data = pd.read_parquet('../data/tables/tbl_merchants.parquet')

# extract the real take rates from companies that do not need imputation
tags = list(merchant_data['tags'])
take_rate = [float(tag[-6:-2]) for tag in tags]
real_take_rate = pd.DataFrame({"merchant_abn":merchant_data.index, "take_rate":take_rate})
real_take_rate

# get clusters and their imputed take rate
cluster_with_take_rate = clusters.merge(cluster_take_rate, on = ['label'], how = 'inner')
cluster_with_take_rate = cluster_with_take_rate.rename(columns={'take_rate':'imputed_take_rate'})

# merge imputed and real take rate (which also has merchant ABN)
merchant_take_rate_full = cluster_with_take_rate.merge(real_take_rate, on='merchant_abn', how='left')

merchant_take_rate_full

# combine the two rates columns using dummy variable trick
merchant_take_rate_full['dummy'] = merchant_take_rate_full['take_rate'].apply(lambda x:1 if x>=0 else 0)
merchant_take_rate_full = merchant_take_rate_full.fillna(0)
merchant_take_rate_full['take_rate'] = merchant_take_rate_full['take_rate'] * merchant_take_rate_full['dummy'] + merchant_take_rate_full['imputed_take_rate']

# then once combined, drop the useless rows
merchant_take_rate_full = merchant_take_rate_full.drop(['dummy', 'imputed_take_rate'], axis=1)

merchant_take_rate_full

# export file
merchant_take_rate_full.to_csv('../data/curated/clusters/output/merchant_take_rate_full.csv', index=False)
