==> notebooks/18_R_clean_occupation.ipynb <==
# Clean up ABS occupation

Purpose:
- clean the raw ABS occupation file into rows = postcodes; columns = occupations; column values = proportions of people doing this occupation in this region

import pandas as pd

# create directory persona if does not exist
import os

dirs_to_create = ['../data/curated/persona', '../data/curated/persona/input', '../data/curated/persona/output']

def create_dirs(dirs_to_create):
    # check if it exists as it makedir will raise an error if it does exist
    for dir_to_create in dirs_to_create:
        if not os.path.exists(dir_to_create):
            os.makedirs(dir_to_create)

create_dirs(dirs_to_create)

def inspect_dataframe(df):
    print("rows, columns:", df.shape)
    return df.head()

# read in data
occupation = pd.read_csv('../data/tables/external/by_postcode/occupation.csv')
# note data has been previously hand-cleaned to take out bad formatting and useless columns for practical time saving purposes

inspect_dataframe(occupation)

# pull out column currently containing noisy postcode 
postcode = occupation['OCCP - 1 Digit Level']

# clean the noisy postcode column to leave only postcode
occupation['postcode'] = [x.split(',')[0] for x in postcode]

# drop unnecessary rows
occupation = occupation.drop(['OCCP - 1 Digit Level', 'Not stated', 'Not applicable', 'Inadequately described'], axis = 1)

inspect_dataframe(occupation)

# previously types were str
occupation['Total'] = occupation['Total'].astype(int)

SKIP_COLS = ['Total', 'postcode'] # columns to be skipped

for col in occupation.columns:
    if col not in SKIP_COLS:
        # previously types were str
        occupation[col] = occupation[col].astype(int)
        # now take the proportion
        occupation[f'{col}_%'] = occupation[col]/occupation['Total'] 
        # drop the column after getting its proportion
        occupation = occupation.drop(col, axis=1)

# drop total
occupation = occupation.drop('Total', axis = 1)

inspect_dataframe(occupation)

# export to csv
occupation.to_csv('../data/curated/persona/input/occupation_cleaned.csv', index = False)
