==> notebooks/20_Y_income_clean.ipynb <==
# Preprocessing for persona data -- income 

import pandas as pd
import numpy as np

## Read in file

income = pd.read_csv('../data/tables/external/by_postcode/income.csv').drop(['Negative income','Not stated','Not applicable'],axis=1)

# change column names
income = income.set_axis(['INCP Total Personal Income (weekly)','0', '75', '225', '350', '450','575','725','900','1125','1375','1625','1875','2500','3500','Total'], axis=1, inplace=False)

## Take mean of income for each postcode

mean_output = []
for i in range(len(income)):
    cols = []
    for j in range(1,len(income.columns)-1):
        cols.append(income.iloc[i,j]*int(income.columns[j]))
    mean_output.append("{:.2f}".format(sum(cols)/income.iloc[i,-1]))


## Create a new dataframe to store all statistics of salary by postcode

final = pd.DataFrame()
final['postcode'] = income['INCP Total Personal Income (weekly)']
final['average_salary'] = mean_output

## Checking if there is average salary that is 0/ nan

final[final['average_salary']=='0.00']

final[final['average_salary'] == "nan"]

### 0.00 values are due to no valid statistic in such postcode
### nan values are due to no salary statistic in such postcode

income[final['average_salary'] == '0.00']

income[final['average_salary']=='nan']

### We decide to replace all nan to 0.00 and keep them

final.average_salary = final.average_salary.str.replace('nan', '0.00')

## Statistics about income data

# to get 5 quantiles of the salary in each postcode
Q0_list = []
Q1_list = []
Q2_list = []
Q3_list = []
Q4_list = []
for i in range(len(income)):
    cols = []
    for j in range(1, len(income.columns)-1):
        for k in range(income.iloc[i,j]):
            cols.append(int(income.columns[j]))
    if all(item == 0 for item in cols):
        Q0_list.append("0.00")
        Q1_list.append("0.00")
        Q2_list.append("0.00")
        Q3_list.append("0.00")
        Q4_list.append("0.00")    
    else:
        Q0_list.append("{:.2f}".format(np.quantile(cols,0.1)))
        Q1_list.append("{:.2f}".format(np.quantile(cols,0.25)))
        Q2_list.append("{:.2f}".format(np.quantile(cols,0.5)))
        Q3_list.append("{:.2f}".format(np.quantile(cols,0.75)))
        Q4_list.append("{:.2f}".format(np.quantile(cols,0.9)))

    

# save the quantiles to the final table
final['median_salary'] = Q2_list
final['Q0_salary'] = Q0_list
final['Q1_salary'] = Q1_list
final['Q3_salary'] = Q3_list
final['Q4_salary'] = Q4_list

# keep postcode numbers only, without state name 
final['postcode'] = final.postcode.str.extract('(\d+)')
# Change the quantiles and mean to float, instead of string
final['average_salary'] = final['average_salary'].astype(float)
final['median_salary'] = final['median_salary'].astype(float)
final['Q0_salary'] = final['Q0_salary'].astype(float)
final['Q1_salary'] = final['Q1_salary'].astype(float)
final['Q3_salary'] = final['Q3_salary'].astype(float)
final['Q4_salary'] = final['Q4_salary'].astype(float)

final

final.to_csv('../data/curated/persona/input/income_cleaned.csv', index= False)
