==> notebooks/21_A_ranking_feature_engineering.ipynb <==
# Creating a ranking feature (Loyalty rate)

# Objective: conduct feature engineering to create a loyalty rate for every merchant

# importing require libraries to run the code in this notebook
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt

# runing spark
spark = (
    SparkSession.builder.appName("preprocessing of taxi data")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)

## Feature Engineering for Ranking

#### 1. Creating features
- Loyality rate <br>
&nbsp;  
Groupby each merchant abn, and calculate:  
    $\text{repeat purchase rate} = \frac {number \ of \ repeat \ customers}{total \ customers}$  
    code implementation  
    - number of repeat customer - can be calculated by the number of re-occuring consumer id
    - total customers - total count of customers  

#### 2. Converting features to scores for ranking


# separating the transaction dataset into training and the full set
transaction_sdf = spark.read.parquet("../data/curated/fraud/output/transactions_withoutfraud")
train_transaction_sdf = transaction_sdf.where(F.col("order_datetime") <= "2022-1-16")

### Conduct aggregation and calculate loyality rate

# group by each merchant and count how many consumers purchased in this merchant
merchant_loyality_rate = transaction_sdf.groupBy("merchant_abn")\
                                        .agg(F.countDistinct("consumer_id").alias("Total_consumers"))
train_merchant_loyality_rate = train_transaction_sdf.groupBy("merchant_abn")\
                                                    .agg(F.countDistinct("consumer_id").alias("Total_consumers"))
merchant_loyality_rate.orderBy("merchant_abn")

# find in each merchant, for each consumer whether they have purchased more than once in this merchant
customer_purchase_counts = transaction_sdf.groupBy("merchant_abn", "consumer_id")\
                                          .agg(F.count("merchant_abn").alias("customer_purchase_count"))
customer_purchase_counts = customer_purchase_counts.where(F.col("customer_purchase_count") > 1)
train_customer_purchase_counts = train_transaction_sdf.groupBy("merchant_abn", "consumer_id")\
                                                      .agg(F.count("merchant_abn").alias("customer_purchase_count"))
train_customer_purchase_counts = train_customer_purchase_counts.where(F.col("customer_purchase_count") > 1)
customer_purchase_counts.orderBy("merchant_abn")

# Count the number of customers that has repeated purchase in each merchant
repeated_customers = customer_purchase_counts.groupBy("merchant_abn")\
                                             .agg(F.count("consumer_id").alias("repeated_customers"))
repeated_customers = repeated_customers.withColumnRenamed("merchant_abn", "merchant_abn_del")

train_repeated_customers = train_customer_purchase_counts.groupBy("merchant_abn")\
                                             .agg(F.count("consumer_id").alias("repeated_customers"))
train_repeated_customers = train_repeated_customers.withColumnRenamed("merchant_abn", "merchant_abn_del")
repeated_customers.orderBy("merchant_abn_del")

# merging the counts of repeated customer and total customer in each merchant together
merchant_loyality_rate = merchant_loyality_rate.join(repeated_customers, 
                                                     merchant_loyality_rate.merchant_abn == repeated_customers.merchant_abn_del, 
                                                     "left").drop("merchant_abn_del")
train_merchant_loyality_rate = train_merchant_loyality_rate.join(train_repeated_customers, 
                                                                train_merchant_loyality_rate.merchant_abn == train_repeated_customers.merchant_abn_del, 
                                                                "left").drop("merchant_abn_del")

# for null values, meaning no purchase ever, or no repeated customer 0 is filled in.
merchant_loyality_rate = merchant_loyality_rate.na.fill(value=0,subset=["repeated_customers"])
train_merchant_loyality_rate = train_merchant_loyality_rate.na.fill(value=0,subset=["repeated_customers"])
merchant_loyality_rate

# calculate purchase rate by dividing repeated customers by total customers
merchant_loyality_rate = merchant_loyality_rate.withColumn("repeated_purchase_rate", 
                                                          (F.col("repeated_customers")) / (F.col("Total_consumers")))
train_merchant_loyality_rate = train_merchant_loyality_rate.withColumn("repeated_purchase_rate", 
                                                          (F.col("repeated_customers")) / (F.col("Total_consumers")))

merchant_loyality_rate.orderBy("repeated_purchase_rate", ascending=False)

# check for null values
merchant_loyality_rate.select([F.count(F.when(F.col(c).contains('None') | \
                                                       F.col(c).contains('NULL') | \
                                                      (F.col(c) == '' ) | \
                                                       F.col(c).isNull() | \
                                                       F.isnan(c), c 
                                                      )).alias(c)
                                                       for c in merchant_loyality_rate.columns])

# grouping data by merchant abn to get standard deviation
agg_trans_sdf = transaction_sdf.groupBy("merchant_abn").agg(F.stddev("dollar_value").alias("SD of merchant revenue"))
agg_trans_sdf = agg_trans_sdf.withColumnRenamed("merchant_abn", "merchant_abn_del")

merchant_loyality_rate_vis = merchant_loyality_rate.join(agg_trans_sdf, 
                                                     merchant_loyality_rate.merchant_abn == agg_trans_sdf.merchant_abn_del, 
                                                     "left").drop("merchant_abn_del")
# plotting standard deviation against repeated purchase rate
merchant_loyality_rate_vis = merchant_loyality_rate_vis.select("merchant_abn", "repeated_purchase_rate", "SD of merchant revenue").toPandas()        
merchant_loyality_rate_vis["SD of merchant revenue"] = np.log10(merchant_loyality_rate_vis["SD of merchant revenue"])
sns.scatterplot(data=merchant_loyality_rate_vis, x="repeated_purchase_rate", y="SD of merchant revenue")
plt.xlabel("Repeated Purchase Rate")
plt.title("Repeated Purchase Rate against SD of merchant revenue")
plt.savefig("../plots/RPRvsSD.png")

# output the file
merchant_loyality_rate.toPandas().to_csv('../data/curated/final_model/input/loyalty_full.csv', index=False)
train_merchant_loyality_rate.toPandas().to_csv('../data/curated/final_model/input/loyalty_train.csv', index=False)
