==> notebooks/22_Y_H_A_persona_score.ipynb <==
# Persona Score Notebook
### **Objective**: This notebook aims to define a score for the general profile of customers of a merchant. The higher the score is, the better the merchant.
### **Overview**: The persona score is based on **our research and knowledge about the BNPL industry**, and it is comprised of several factors such as median salary and age distributions of that region. Since we only have the geographical location of the customers, we will take the mean of all customers. <u>See the codes for more details</u>.

from pyspark.sql import SparkSession
from pyspark.sql import functions as F
import pandas as pd
import numpy as np
from functools import reduce

spark = (
    SparkSession.builder.appName("preprocessing of taxi data")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "15g")
    .getOrCreate()
)

age_sdf = spark.read.option("header", True).csv("../data/tables/external/by_postcode/1 year age.csv")
education_sdf = spark.read.option("header", True).csv("../data/tables/external/by_postcode/education.csv")

columns = ["sub_total", "18_24", "25_34", "35_44", "45_54", "55_64", "65_plus", "postcode"]
age_group = columns[:-1]
age_sdf = age_sdf.withColumn(columns[0], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(18, 116)])) \
                .withColumn(columns[1], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(18, 25)]) / F.col(columns[0])) \
                .withColumn(columns[2], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(25, 35)]) / F.col(columns[0])) \
                .withColumn(columns[3], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(35, 45)]) / F.col(columns[0])) \
                .withColumn(columns[4], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(45, 55)]) / F.col(columns[0])) \
                .withColumn(columns[5], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(55, 65)]) / F.col(columns[0])) \
                .withColumn(columns[6], reduce(lambda x,y:x+y, [F.col(str(x)) for x in range(65, 116)]) / F.col(columns[0])) \
                .withColumn(columns[7], F.regexp_extract('AGEP Age', r'\d+', 0)) \
                .select(*columns)
age_sdf.limit(5)

student_sdf = education_sdf.withColumn("postcode", F.regexp_extract("postcode", r'\d+', 0)) \
                           .withColumnRenamed("total", "total_parttime_or_fulltime_students") \
                           .select("postcode", "total_parttime_or_fulltime_students")

student_sdf.limit(5)

income_sdf = spark.read.option("header", True).csv("../data/curated/persona/input/income_cleaned.csv") \
                                              .select("postcode", "median_salary")
income_sdf.limit(5)

def mean(column_name):
   return F.mean(column_name).alias("mean_" + column_name)

def preprocessing_pipeline(dlist):
    """
    preprocess all dataframe in the list
        parameters: 
                   a list of dataframes
        returns:
                a generator of preprocessed dataframes
    """
    for data in dlist:
        # join persona data
        data = data.join(income_sdf, "postcode", "left") \
                   .join(student_sdf, "postcode", "left") \
                   .join(age_sdf, "postcode", "left")

        # remove rows contain null in persona
        data = data.filter(F.col("median_salary").isNotNull() \
                         | F.col("sub_total").isNotNull() \
                         | F.col("total_parttime_or_fulltime_students").isNotNull())

        # take the mean of all persona
        agg_data = data.groupBy("merchant_abn") \
                       .agg(
                            mean("median_salary"),
                            mean("total_parttime_or_fulltime_students"),
                            *[mean(name) for name in age_group]
                            )
        yield agg_data
        

transaction_sdf = spark.read.parquet("../data/curated/fraud/output/transactions_withoutfraud") \
                            .drop('user_id2',
                                  'order_datetime2',
                                  'fraud rate',
                                  'fraud2',
                                  'user_id3',
                                  'order_datetime3',
                                  'fraud_probability',
                                  'fraud3')

# training data is the part of data before the first fortnight of 2022
train_transaction_sdf = transaction_sdf.where(F.col("order_datetime") <= "2022-1-16")

data_list = [transaction_sdf, train_transaction_sdf]

data_list = list(preprocessing_pipeline(data_list))

data_list[0].limit(5)

def compute_persona_score(dlist, filenames):
    """
    compute persona scores for all merchants in a dataframe and write the result to a .csv file
    parameters:
               dlist (list): a list of dataframes
               filenames (list): a list of filenames in which the final results will be written
    return:
           a generator of dataframes containing merchant ABN and persona score
    """
    for data, filename in zip(dlist, filenames):
        data = data.withColumn("pop_index",
                                0.26*F.col("mean_18_24") + 
                                0.35*F.col("mean_25_34") + 
                                0.20*F.col("mean_35_44") + 
                                0.12*F.col("mean_45_54") + 
                                0.05*F.col("mean_55_64") +
                                0.01*F.col("mean_65_plus")
                                )
                                
        data = data.select("merchant_abn",
                            "mean_median_salary", 
                            "mean_total_parttime_or_fulltime_students",
                            "pop_index",
                            "mean_sub_total")

        for col in data.columns[1:]:
            col_max = data.agg({col:"max"}).collect()[0][0]
            col_min = data.agg({col:"min"}).collect()[0][0]
            data = data.withColumn(col, (F.col(col)-F.lit(col_min)) / (F.lit(col_max)-F.lit(col_min)) )
            data = data.withColumnRenamed(col, "normalized_"+col)
        

        data = data.withColumn("persona_score",
                               F.col("normalized_mean_median_salary") *
                               F.col("normalized_mean_total_parttime_or_fulltime_students") *
                               F.col("normalized_pop_index") *
                               F.col("normalized_mean_sub_total")
                               )
        
        data.select("merchant_abn", "persona_score").toPandas().to_csv("../data/curated/final_model/input/" + filename, index=False)
        
        print("%s is done" % filename)

        yield data


filenames = ["persona_full.csv", "persona_train.csv"]
list(compute_persona_score(data_list, filenames))[0].limit(5)
